remove pre-refactor code
