+++ megaman/geometry/adjacency.py
