+++ megaman/geometry/tests/test_adjacency.py
@@ -5,7 +5,7 @@ from numpy.testing import assert_allclose, assert_raises
 from scipy.sparse import isspmatrix
 from scipy.spatial.distance import cdist, pdist, squareform
 
-from megaman.geometry.adjacency_new import compute_adjacency_matrix, Adjacency
+from megaman.geometry.adjacency import compute_adjacency_matrix, Adjacency
 
 
 try:
