+++ megaman/geometry/tests/test_laplacian.py
@@ -6,9 +6,9 @@ from numpy.testing import assert_allclose, assert_equal, assert_raises
 from scipy.sparse import isspmatrix, csr_matrix
 from scipy import io
 
-from megaman.geometry.adjacency_new import compute_adjacency_matrix
-from megaman.geometry.affinity_new import compute_affinity_matrix
-from megaman.geometry.laplacian_new import Laplacian, compute_laplacian_matrix
+from megaman.geometry.adjacency import compute_adjacency_matrix
+from megaman.geometry.affinity import compute_affinity_matrix
+from megaman.geometry.laplacian import Laplacian, compute_laplacian_matrix
 
 
 TEST_DATA = os.path.join(os.path.dirname(__file__),
